+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [1.7.2](https://github.com/luoxuhai/pcl.js/compare/v1.7.1...v1.7.2) (2022-10-01)
+
+
+### 🐛 Bug Fixes
+
+* can't get PointXYZRGB to rgb value ([be97496](https://github.com/luoxuhai/pcl.js/commit/be974960bd4d84085b4a5db2e3bf4cb83b3abfb8))
+
+
+### ✨ Features
+
+* new readPCDHeader method ([2b56103](https://github.com/luoxuhai/pcl.js/commit/2b56103df15740f7a36b9a4144b334ebe931ba0d))
+
+## [1.7.1](https://github.com/luoxuhai/pcl.js/compare/v1.7.0...v1.7.1) (2022-09-29)
+
+
+### 🐛 Bug Fixes
+
+* get points cause memory leak ([7bbc015](https://github.com/luoxuhai/pcl.js/commit/7bbc01546f963006bdfe1e2f856266cd05884c1c))
+* points.size is not a function at PointCloudViewer.addPointCloud ([fea411f](https://github.com/luoxuhai/pcl.js/commit/fea411f42e424a0e18c38f25af6c82b09fda7ffe))
+
 ## [1.7.0](https://github.com/luoxuhai/pcl.js/compare/v1.6.0...v1.7.0) (2022-09-29)